Moved polish notation IO into separate files and changed makefile
