+++ 02.10.2025/Makefile
@@ -1,7 +1,7 @@
 BUILDDIR:=build
 SRC:=$(wildcard *.c)
 SRC:=$(filter-out stack.c, $(SRC))
-OUT:=$(patsubst %.c, $(BUILDDIR)/%.out, $(SRC))
+OUT:=$(BUILDDIR)/stack.o $(BUILDDIR)/polishNotation.out
 
 CFLAGS=-Wall -pedantic -Wextra
 CC=gcc
@@ -14,8 +14,8 @@ $(BUILDDIR):
 $(BUILDDIR)/stack.o: stack.h stack.c
 	$(CC) $(CFLAGS) -c stack.c -o $(BUILDDIR)/stack.o
 
-$(BUILDDIR)/%.out: %.c
-	$(CC) $(CFLAGS) -o $@ $^ $(BUILDDIR)/stack.o
+$(BUILDDIR)/polishNotation.out: polishNotation.c polishNotationIO.c polishNotationIO.h $(BUILDDIR)/stack.o
+	$(CC) $(CFLAGS) -o $@ polishNotation.c polishNotationIO.c $(BUILDDIR)/stack.o
 
 clean:
 	rm -rf $(BUILDDIR)
