+++ 02.10.2025/polishNotation.c
@@ -1,4 +1,5 @@
 #include "stack.h"
+#include "polishNotationIO.h"
 #include <stdio.h>
 #include <ctype.h>
 
@@ -9,51 +10,6 @@
     fprintf(stderr, __VA_ARGS__); \
     *pExitCode = 1; \
 } while(0)
-
-static int _putback = '\0';
-static char _outputBuf[BUFSIZ];
-static int _outputBufSize = 0;
-/*
- * Skip white spaces and tabs and return character from the input.
- */
-int getInput()
-{
-    int ch;
-    if (_putback) {
-        ch = _putback;
-        _putback = '\0';
-        return ch;
-    }
-    for (ch = getchar(); isblank(ch); ch = getchar());
-    return ch;
-}
-
-void putOutput(int ch)
-{
-    if (_outputBufSize < (int)(sizeof(_outputBuf) - 2)) {
-        _outputBuf[_outputBufSize++] = ch;
-        _outputBuf[_outputBufSize++] = ' ';
-    }
-}
-
-void printOutput()
-{
-    _outputBuf[_outputBufSize] = '\0';
-    printf("%s\n", _outputBuf);
-}
-
-/*
- * Puts character back into the stream.
- */
-void putback(int ch) 
-{
-    _putback = ch;
-}
-
-bool isEOF(int ch)
-{
-    return ch == EOF || ch == '\n';
-}
 
 /*
  * Get operation precedence in expression.
